pyproject.toml
[tool.poetry]
name = "mb"
version = "0.1.0"
description = "MessageBoard backend written in FastAPI"
authors = ["Gleb Voropaev <info@gvoropaev.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.29.0"
fastapi = "^0.110.1"
pyjwt = "^2.8.0"
sqids = "^0.4.1"
pydantic-settings = "^2.2.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.0"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
